refactor(cockpit): extract mutation gate module
